requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
hedera-sdk-py==2.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyjnius==1.5.0
requests==2.31.0
urllib3==2.0.2
Werkzeug==2.3.4
